Fixing Regex
The regex has been catching "+"; "$"; "^"; "|"; "~" and "`", really annoying issue. I might be dumb and catching those might have an actual use, but as far as I can tell it's nothing more than an annoying issue.
Also I'm sure there is a better way of going about it, but I'm tired and cannot be bothered to do regex.